<Unit Test>[Commands]: Write unit tests for serialization
Deserialization unit tests come next

[https://github.com/hasii2011/PyUt/issues/273]

+++ src/org/pyut/miniogl/SelectAnchorPoint.py
@@ -15,7 +15,7 @@ class SelectAnchorPoint(AnchorPoint, ShapeEventHandler):
     to attach something
 
     """
-    def __init__(self, x: float, y: float, attachmentPoint: AttachmentPoint, parent: Shape = None):
+    def __init__(self, x: int, y: int, attachmentPoint: AttachmentPoint, parent: Shape = None):
         """
 
         Args:
